feat: add UserRepository and implement login functionality using Koin.
This commit introduces a `UserRepository` and a mock implementation `MockUserRepository` for user data management.
It also adds the `UserRepository` to the Koin dependency injection setup in `AppModule.kt`.
The `LoginScreen` now utilizes Koin to access the `UserRepository`, and implements login functionality by calling `UserRepository.loginUser()`.
Additionally, the UI of the login screen now reflects the name of logged user.